service/sso_gateway/pyproject.toml
[tool.poetry]
name = "sso-gateway"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
uvicorn = {extras = ["standard"], version = "^0.34.0"}
gunicorn = "^23.0.0"
grpcio = "^1.69.0"
protobuf = "^5.29.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
